Re-implemented compute_stats function
Re-implemented compute_stats() function because the code generated by ChatGPT was raising errors, and it was not computing what we asked (cumulative statistics in the statistics table).

This version returns, for each motif (group), tue AuROC, the AuPR and a stat table with cumulative statistics (TP, FP, TN, FN, FPR, TPR, PPV) as a function of decreasing scores.

The computed values for AuROC and AuPR are not identical by consistent with those returned by sklearn.metrics, up to the 3 decimals. The code could be improved.