Added suffix to ID of mutated matrices
In Transfac format, each matrix has an accession (AC) and an identifier (ID). The rationale is that in the Transfac database, each PSSM had a user-friendly identifier (indicating the taxon + the name of the transcription factor + an index number) and a non-interpretable accession (supposedly assigned by the DBMS).

For matrices generated by RSAT (e.g. resulting from motif discovery), the ID and AC are identical since there is no associated databaase.

In the previous versions, the mutant suffix "_G{gen_nb}_M{matrix_nb}_C{child_nb}" was appended only to the AC. In this version we append it to both AC and ID to ensure consistency.