Added a second study case
Added LEF1 HTS as a second study case in addition to GABPA CHS. This study case was chosen because the initial AuROC are weak (between 0.376 and 0.778) so we can evaluate the improvement.

Note: a strange effect: in gen1 all the AuROC scores in the transfac file range from 0.855 to 0.888, although in principle they should keep the score of the generation 0 in the history tracking comments. There is obviously a bug when re-writing the comments.