Fixed bugs, working version
Fixed some bugs with the naming of the files to store.
The json and CSV export are commented because they relied on a re-running of score_matrices. I should add a way for `genetic_algorithm()` to collect a data.frame with the different matrix scores at each generation, and return it at the end together with the final matrices.