Created unique tmp dir
Adapted the temporary directory to avoid clashes between concurrent runs (e.g. via slurm job scheduler of a cluster). The temporary directory is now a sub-directory of the output directory, with a specific sub-subdirectory with the date and time of the run.

Also revised the level of verbosity for all the log messages.