ChatGPT-generated code after iterative prompting
- added many options to fine-tune the graph
- added verbosity
- export the plot to files
- did not restore the description of file format that I previously asked to insert and which was previously  displayed correctly in the help message.

The result is really nice,  although there are some problem that ChatGPT did not solve even after repeated requests
- `--y_min` is not working correctly, the Y axis limit is 0.1 below what I ask
- the subtitle is displayed above the title

Prompt:

1. Could you add parameters --min_x and --max_x to specify the min and max values for the  Y axis? If not specified , the range of the X axis will be set automatically as is currently the case, but if either of them or both are specified, they will specifically set the upper and/or lower limits of the X axis.

2. Sorry, I meant the Y axis, not the X axis

3. Great. I would add two requests:

- horizontal grid should be slightly more detailed. In particular, if I display AUROC scores (Y) from 0 to 1, I would like to have one bar every 0.1 step.

- I would also like to add two optional parameters --min_g and --max_g, which would restrict the whole analysis to this range of generations. Importantly, the range of AuROC values (Y values) displayed by default (when options --min_y and --max_y are not invoked) should be adapted to take into account only the values present in the considered range of generations

4. Could you also add an optional parameter --y_step1 and --y_step2 to specify the steps of a first level (mid-gray) and second level (pale grey) of steps in the horizontal grid?

5. Could you also print a log message on the terminal console with the complete command used to generate the graph?

6. and associate it to a parameter -v/--verbosity: if higher than 0, the message will be displayed

7 Bug report:

````
ValueError: 'midgray' is not a valid value for color: supported inputs are (r, g, b) and (r, g, b, a) 0-1 float tuples; '#rrggbb', '#rrggbbaa', '#rgb', '#rgba' strings; named color strings; string reprs of 0-1 floats for grayscale values; 'C0', 'C1', ... strings for colors of the color cycle; and pairs combining one of the above with an alpha value
```

Can you use these colors:  for midgray: '#888888' and for pale gray '#CCCCCC' ?

8. The --min_g and --maxG parameters create a bug.
 [... error code pasted...]

9. I don't think that this was the problem. First, I do have data in the generations from 10 to 20, and secondly the new code still bugs .
 [...error code pasted...]

10. Great. Still some detail: I used the following parameters:
--min_g 10 --max_g 20 --min_y 0.78 --max_y 0.92

The generations are well set, but the Y axis sgoes from 0.7 to 0.92.

11. Can you add parameters --xsize and --ysize in inches for the exported plot ?
It woudl also be useful to have an option  --dots_per_inch for the png output, but I don't know if this is difficult.

12. There is really a problem with --min_y. I tried with the following options:

-i data/score_tables/PRDM5_GHTS_YWK_B_AffSeq_B1_PRDM5.C2_clust-trimmed-matrices_train-vs-rand_gen0-20_score_table.tsv -o data/score_tables/PRDM5_GHTS_YWK_B_AffSeq_B1_PRDM5.C2_clust-trimmed-matrices_train-vs-rand_gen0-20_score_table_plot.pdf -f pdf -t "PRDM5_GHTS_YWK_B_AffSeq_B1_PRDM5.C2_clust-trimmed-matrices " -s "train versus rand" --y_step1 0.1 --y_step2 0.01 --min_y 0.5 --max_y 1 --xsize 16 --ysize 8

I have a graph from 0.4 to 1.0 instead of 0.5 to 1.0. The grid is only displayed between 0.5 and 1.0, which leaves a white area between 0.4 and 0.5. Can you please check ?

For the usage example in argparse doc, the name of the program should be replaced by plot-auroc-profiles (I think this name is more appropriate).

13. Problem not solved, but OK let us take it as such, I will try to debug it with human brain (no  offense I hope, what you did is already great).

I noticed that you forgot the description of the input file format (column contents) that was present in  an earlier version. Can you restore it ?

Note: I asked ChatGPT to list the last 10, 15 or 20 promtps. It enumerates prompts, but the list is not exhaustive, and when I increase the  number of prompts requested it repeats the same prompt rather than telling me that there were less than 20 prompts ;-)